For #16395 - Improve contrast ratio on Site Permissions subtext
Right now, the subtext in SitePermissionsManagePhoneFeatureFragment always
displays as Color.gray (#888888).

Against the default light theme background (#F9F9FB), the contrast ratio
is ~3.37. For accessibility purposes, it would be preferable to use colors
that result in a contrast ratio greater than 4.50 for small text.

Let's resolve this issue by migrating to the following color schemes:

| Theme Name | Background Color | Foreground Color | Contrast Ratio |
|------------|------------------|------------------|----------------|
| Light      | #F9F9FB          | #5B5B66          |           6.37 |
| Dark       | #1C1B22          | #BFBFC9          |           9.36 |